requirements.txt
discord
python-dotenv
mongoengine